# app/govern/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  bll.AddRoleReq:
    properties:
      enable:
        type: boolean
      group:
        type: string
      icon:
        type: string
      memo:
        type: string
      name:
        type: string
      seq:
        type: integer
    required:
    - name
    - seq
    type: object
...
  bll.AuthorizeReq:
    properties:
      menuWidgetMap:
        additionalProperties:
          items:
            type: string
          type: array
        type: object
    required:
    - menuWidgetMap
    type: object
...
  bll.EnableRoleReq:
    properties:
      enable:
        type: boolean
    required:
    - enable
    type: object
...
  bll.ListRoleResp:
    properties:
      data:
        type: object
      total:
        type: integer
    type: object
...
  bll.ProfileRoleResp:
    properties:
      createdAt:
        type: integer
      creator:
        type: string
      enable:
        type: boolean
      group:
        type: string
      icon:
        type: string
      memo:
        type: string
      name:
        type: string
      seq:
        type: integer
      updatedAt:
        type: integer
    type: object
...
  bll.UpdateRoleReq:
    properties:
      group:
        type: string
      icon:
        type: string
      memo:
        type: string
      name:
        type: string
      seq:
        type: integer
    type: object
...
host: localhost:8000
info:
  contact:
    email: sfshf@github.com
    name: gavin
    url: http://github.com/sfshf
  description: This is a back-end government app.
  license:
    name: MIT
    url: https://github.com/sfshf/sprout/blob/main/LICENSE
  termsOfService: http://swagger.io/terms/
  title: Govern APIs
  version: 0.0.1-beta
paths:
...
  /roles:
    get:
      description: Get a list of role.
      operationId: role-list
      parameters:
      - in: query
        name: createdAtBegin
        type: integer
      - in: query
        name: createdAtEnd
        type: integer
      - in: query
        name: creator
        type: string
      - in: query
        name: enable
        type: boolean
      - in: query
        name: group
        type: string
      - in: query
        name: name
        type: string
      - in: query
        name: orderBy
        type: string
      - in: query
        name: page
        type: integer
      - in: query
        name: perPage
        type: integer
      - in: query
        name: seq
        type: integer
      produces:
      - application/json
      responses:
        "1000":
          description: feasible and predictable errors.
          schema:
            type: error
        "2000":
          description: role list.
          schema:
            $ref: '#/definitions/bll.ListRoleResp'
      security:
      - ApiKeyAuth: []
      summary: Get a list of role.
      tags:
      - role
    post:
      consumes:
      - application/json
      description: Add a new role.
      operationId: role-add
      parameters:
      - description: required attributes to add a new role.
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/bll.AddRoleReq'
      produces:
      - application/json
      responses:
        "1000":
          description: feasible and predictable errors.
          schema:
            type: error
        "2000":
          description: successful action.
          schema:
            type: "null"
      security:
      - ApiKeyAuth: []
      summary: Add a new role.
      tags:
      - role
  /roles/:id:
    delete:
      description: Evict a specific role.
      operationId: role-evict
      parameters:
      - description: id of the role to evict.
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "1000":
          description: feasible and predictable errors.
          schema:
            type: error
        "2000":
          description: successful action
          schema:
            type: "null"
      security:
      - ApiKeyAuth: []
      summary: Evict a specific role.
      tags:
      - role
    get:
      description: Get the profile of a role.
      operationId: role-profile
      parameters:
      - description: id of the role.
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "1000":
          description: feasible and predictable errors.
          schema:
            type: error
        "2000":
          description: profile of the role.
          schema:
            $ref: '#/definitions/bll.ProfileRoleResp'
      security:
      - ApiKeyAuth: []
      summary: Get infos of a role.
      tags:
      - role
    put:
      consumes:
      - application/json
      description: Update a specific role.
      operationId: role-update
      parameters:
      - description: id of the role to evict.
        in: path
        name: id
        required: true
        type: string
      - description: attributes need to update.
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/bll.UpdateRoleReq'
      produces:
      - application/json
      responses:
        "1000":
          description: feasible and predictable errors.
          schema:
            type: error
        "2000":
          description: successful action.
          schema:
            type: "null"
      security:
      - ApiKeyAuth: []
      summary: Update a specific role.
      tags:
      - role
  /roles/:id/authorize:
    put:
      consumes:
      - application/json
      description: Allocate authorities to a specific role using menu-widgets pairs.
      operationId: role-allocate-authority
      parameters:
      - description: id of the role to be allocated authorities.
        in: path
        name: id
        required: true
        type: string
      - description: menu-widgets pairs.
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/bll.AuthorizeReq'
      produces:
      - application/json
      responses:
        "1000":
          description: feasible and predictable errors.
          schema:
            type: error
        "2000":
          description: successful action.
          schema:
            type: "null"
      security:
      - ApiKeyAuth: []
      summary: Allocate authorities to a specific role.
      tags:
      - role
  /roles/:id/enable:
    patch:
      consumes:
      - application/json
      description: Enable or disable a role.
      operationId: role-enable
      parameters:
      - description: id of the role.
        in: path
        name: id
        required: true
        type: string
      - description: true for enable, or false for disable.
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/bll.EnableRoleReq'
      produces:
      - application/json
      responses:
        "1000":
          description: feasible and predictable errors.
          schema:
            type: error
        "2000":
          description: successful action.
          schema:
            type: "null"
      security:
      - ApiKeyAuth: []
      summary: Enable or disable a role.
      tags:
      - role
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"